feature:#5 로테이션 페이지 보완 및 빌드 오류 수정
