+++ .env
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_RIOT_API_KEY="RGAPI-f97c9228-77c5-4b1b-8ebb-d46ebafa4147"
+NEXT_PUBLIC_RIOT_API_KEY="RGAPI-8fe7f901-31e4-4390-9a4a-5f5429131a0d"
