+++ src/app/loading.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Loading = () => {
+  return <div>로딩 중...</div>;
+};
+
+export default Loading;
